+++ CHANGELOG.md
@@ -2,8 +2,9 @@ Changelog
 ===
 
 ### 18.10.2022
-- Updated startpage gui
+- Updated start page gui
 - Updated taxonomy browser
+- Added easy taxonomy identifier lookup
 
 ### 14.10.2022
 - Added new Metadata options selections